refactor: correct barrel files
